Move humanizeEstimatedTime to Format Utils
